backend/pyproject.toml
[project]
name = "legartis-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "flask>=3.1.2",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
